+++ src/lib/google-analytics/index.tsx
@@ -9,13 +9,15 @@ export default function GoogleAnalytics({ gaId }: { gaId: string }) {
       />
       <Script
         id="google-analytics"
+        strategy="afterInteractive"
         dangerouslySetInnerHTML={{
           __html: `
             window.dataLayer = window.dataLayer || [];
             function gtag(){dataLayer.push(arguments);}
             gtag('js', new Date());
-
-            gtag('config', '${gaId}');
+            gtag('config', '${gaId}', {
+              page_path: window.location.pathname,
+            });
           `,
         }}
       />
